Major code refactorization (work in progress).
This is the first commit of the new release.
Now "core" methods are split into separate package, and it can
be used by code that provides user interface (or whatever).

Provided script (bin/querypkg) is somewhat functional; it can query
a user (the same way as the previous version), get and display
the results in a simple way (this is to be improved).

Functional improvements in comparison with the previous version:
- `:' is now allowed in query; it turns out it wouldn't be
  accidentally interpreted as an API option, so it's fine (one can
  query SLOTs now with "match" search type :o),
- don't bail out immediately if more than one keyword provided;
  it should provide more useful feedback in such cases;
- bigger timeout with "path" search type,
- interactive mode has colours now!